requirements_aqt.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
cplex==22.1.1.2
dill==0.3.9
docplex==2.29.241
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
pbr==6.1.0
platformdirs==4.3.6
psutil==6.1.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qiskit==1.3.2
qiskit-aer==0.16.0
qiskit-algorithms==0.3.1
qiskit-aqt-provider==1.10.0
qiskit-optimization==0.6.1
rustworkx==0.15.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
stevedore==5.4.0
symengine==0.13.0
sympy==1.13.3
tabulate==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
